requirements.txt
# Core dependencies
requests>=2.31.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Web scraping and browser automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Database and vector operations
sqlite-vec>=0.1.0
sentence-transformers>=2.2.0

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0
click>=8.1.0
